Restructured config.yml authorization -> userConfig
