+++ src/main/java/nl/knaw/dans/sword2/config/DefaultUserConfig.java
@@ -1,0 +1,4 @@
+package nl.knaw.dans.sword2.config;
+
+public class DefaultUserConfig {
+}
